Add support for restarting bhyve on reboot.
The design of bhyve(8) is that the process exits when a virtual machine
reboots, but packer needs the machine to perform the first boot so that
it can finish the provision over SSH.

Watch for bhyve exits and restart using the same arguments if the exit
status is 0 (guest reboot).

We don't yet correctly propagate any errors up to packer if the restart
fails, but do log it.