Support file-backed disks.
This is now the default.  Simplify zvol support by introducing a
disk_use_zvol variable which defaults to false.  If set then will enable
zvol support, and disk_zpool (renamed from "zpool") controls which zpool
the zvol will be created on, and defaults to "zones".

These changes mean configuration is simpler, should work out of the box,
and uses consistent variable naming in line with qemu.